Use custom DNS server without adding it to system
Currently, SIVT will fail if the DNS resolver it was configured with cannot
resolve custom Avi FQDNs. This might not work for customers who want to
use FQDNs that will only live on the DNS resolver that they provided to the
UI.

This commit fixes that by using `dns.resolver` and monkey-patching
`urllib.connection.create_connection` to ensure that the resolver that
the user provided is used during environment turn up.

Signed-off-by: Carlos Nunez <ncarlos@vmware.com>